# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dots-lite bridge",
    "version": "1"
  },
  "paths": {
...
    "/api/device": {
      "get": {
        "summary": "Device",
        "operationId": "device_api_device_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/projects": {
      "get": {
        "summary": "Project List",
        "operationId": "project_list_api_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "type": "array",
                  "title": "Response Project List Api Projects Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/agents": {
      "get": {
        "summary": "Agent List",
        "operationId": "agent_list_api_agents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentResponse"
                  },
                  "type": "array",
                  "title": "Response Agent List Api Agents Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/tasks": {
      "post": {
        "summary": "Create Task",
        "operationId": "create_task_api_tasks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Tasks",
        "operationId": "list_tasks_api_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "summary": "Task Detail",
        "operationId": "task_detail_api_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/cancel": {
      "post": {
        "summary": "Cancel Task",
        "operationId": "cancel_task_api_tasks__task_id__cancel_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/messages": {
      "post": {
        "summary": "Followup",
        "operationId": "followup_api_tasks__task_id__messages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowupRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "supports_resume": {
            "type": "boolean",
            "title": "Supports Resume"
          }
        },
        "type": "object",
        "required": [
          "id",
          "available",
          "supports_resume"
        ],
        "title": "AgentResponse"
      },
      "Capability": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          }
        },
        "type": "object",
        "required": [
          "available"
        ],
        "title": "Capability"
      },
      "CreateTaskRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Project Id"
          },
          "prompt": {
            "type": "string",
            "maxLength": 32000,
            "minLength": 1,
            "title": "Prompt"
          },
          "agent_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Agent Id",
            "default": "codex"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project_id",
          "prompt"
        ],
        "title": "CreateTaskRequest"
      },
      "DeviceResponse": {
        "properties": {
          "api_version": {
            "type": "integer",
            "title": "Api Version"
          },
          "instance_id": {
            "type": "string",
            "title": "Instance Id"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "hostname": {
            "type": "string",
            "title": "Hostname"
          },
          "status": {
            "type": "string",
            "const": "online",
            "title": "Status"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "capabilities": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Capability"
            },
            "type": "object",
            "title": "Capabilities"
          }
        },
        "type": "object",
        "required": [
          "api_version",
          "instance_id",
          "id",
          "name",
          "hostname",
          "status",
          "platform",
          "version",
          "capabilities"
        ],
        "title": "DeviceResponse"
      },
      "FollowupRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 32000,
            "minLength": 1,
            "title": "Prompt"
          },
          "agent_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "FollowupRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ProjectResponse"
      },
      "TaskDetail": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "parent_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Task Id"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "output": {
            "type": "string",
            "title": "Output"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "output_truncated": {
            "type": "boolean",
            "title": "Output Truncated"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "project_id",
          "agent_id",
          "status",
          "created_at",
          "started_at",
          "finished_at",
          "exit_code",
          "session_id",
          "parent_task_id",
          "conversation_id",
          "prompt",
          "output",
          "error",
          "output_truncated"
        ],
        "title": "TaskDetail"
      },
      "TaskPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "offset",
          "limit",
          "total"
        ],
        "title": "TaskPage"
      },
      "TaskSummary": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "parent_task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Task Id"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "project_id",
          "agent_id",
          "status",
          "created_at",
          "started_at",
          "finished_at",
          "exit_code",
          "session_id",
          "parent_task_id",
          "conversation_id"
        ],
        "title": "TaskSummary"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}